+++ database/seeders/DatabaseSeeder.php
@@ -1,5 +1,8 @@
 <?php
 
+namespace Database\Seeders;
+
+use Database\Seeders\CharacterSeeder;
 use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
