+++ database/seeders/PatchSeeder.php
@@ -1,8 +1,10 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
-class ModStatSeeder extends Seeder
+class PatchSeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -11,8 +13,6 @@ class ModStatSeeder extends Seeder
      */
     public function run()
     {
-        $sql = file_get_contents(database_path().'/seeds/mods_stats_import.sql');
-
-        DB::statement($sql);
+        \App\Patch::factory()->count(20)->create();
     }
 }
