+++ database/seeders/PostSeeder.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Seeders;
+
 use App\Post;
 use Illuminate\Database\Seeder;
