--- database/seeds/PatchSeeder.php
@@ -1,16 +1,0 @@
-<?php
-
-use Illuminate\Database\Seeder;
-
-class PatchSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        \App\Patch::factory()->count(20)->create();
-    }
-}
